Renamed directory to /src/main/java/com/acmcsuf/crying_counter/
