+++ src/main/java/com/acmcsuf/crying_counter/Bot.java
@@ -1,4 +1,4 @@
-package com.acmcsuf.bot_committee;
+package com.acmcsuf.crying_counter;
 
 import java.util.Arrays;
 import java.util.List;
